+++ test/unit/test-bench.js
@@ -55,9 +55,9 @@ function benchmark(testName, count, callback) {
 //    = profileWrapper($.ui.fancytree._FancytreeNodeClass.prototype.addChildren);
 
 function addNodes(node, level1, level2, level3, forceUpdate) {
-//	if( forceUpdate !== true ){
-//		node.tree.enableUpdate(false);
-//	}
+	if( forceUpdate !== true ){
+		node.tree.enableUpdate(false);
+	}
 	var d, f, i, j, k, key;
 	for(i=0; i<level1; i++) {
 		key = "" + (i+1);
@@ -78,7 +78,9 @@ function addNodes(node, level1, level2, level3, forceUpdate) {
 			}
 		}
 	}
-//	node.tree.enableUpdate(true);
+	if( forceUpdate !== true ){
+		node.tree.enableUpdate(true);
+	}
 }
 
 
@@ -86,7 +88,7 @@ function addNodes(node, level1, level2, level3, forceUpdate) {
 
 module("API");
 
-test("Trigger events", function() {
+test("Trigger events", function(assert) {
 	expect(3);
 	benchmark("10000 createNode DOM events (widget._trigger())", 10000, function() {
 		var i,
@@ -125,7 +127,7 @@ test("Trigger events", function() {
 
 module("Standard tree");
 
-test("Add nodes using API to collapsed node (no rendering)", function() {
+test("Add nodes using API to collapsed node (no rendering)", function(assert) {
 	expect(3);
 
 	var tree = _resetEmptyTree();
@@ -149,7 +151,7 @@ test("Add nodes using API to collapsed node (no rendering)", function() {
 });
 
 
-test("Create and render hidden nodes, but don't make visible (i.e. don't expand)", function() {
+test("Create and render hidden nodes, but don't make visible (i.e. don't expand)", function(assert) {
 	expect(3);
 
 	var tree = _resetEmptyTree();
@@ -175,11 +177,12 @@ test("Create and render hidden nodes, but don't make visible (i.e. don't expand)
 });
 
 
-test("Expand 1000 nodes deep (with 10 top level nodes, triggers expand -> render and display)", function() {
+test("Expand 1000 nodes deep (with 10 top level nodes, triggers expand -> render and display)", function(assert) {
 	expect(5);
 
 	var tree = _resetEmptyTree(),
 		node = tree.getNodeByKey("root"),
+		done = assert.async(),
 		timer = new tools.AsyncTimer("1000 deep (10x10x10) with expand", 1000);
 
 	addNodes(node, 10, 10, 10);
@@ -188,22 +191,25 @@ test("Expand 1000 nodes deep (with 10 top level nodes, triggers expand -> render
 	node.setExpanded().done(function(){
 		timer.subtime("expand done");
 		// force browser to re-flow?
-//		var dummy = tree.$div[0].offsetHeight;
+		//jshint unused:false
+		var dummy = tree.$div[0].offsetHeight;
 		timer.subtime("calc offsetHeigth, (force reflow?)");
 		setTimeout(function(){
 			// yield, so browser can redraw
 			timer.subtime("setTimeout(0) to allow reflow/redraw");
 			timer.stop();
+			done();
 		}, 0);
 	});
 });
 
 
-test("Expand 1000 top level nodes (triggers expand -> render and display)", function() {
+test("Expand 1000 top level nodes (triggers expand -> render and display)", function(assert) {
 	expect(2);
 
 	var tree = _resetEmptyTree(),
 		node = tree.getNodeByKey("root"),
+		done = assert.async(),
 		timer = new tools.AsyncTimer("1000 top level nodes flat with expand", 1000);
 
 	addNodes(node, 1000, 0, 0);
@@ -211,17 +217,19 @@ test("Expand 1000 top level nodes (triggers expand -> render and display)", func
 
 	node.setExpanded().done(function(){
 		timer.stop();
+		done();
 	});
 });
 
 
-test("Expand 1000 top level nodes with ARIA and checkboxes (triggers expand -> render and display)", function() {
+test("Expand 1000 top level nodes with ARIA and checkboxes (triggers expand -> render and display)", function(assert) {
 	expect(2);
 
 	var tree = _resetEmptyTree({
 			aria: true,
 			checkbox:true
 		}),
+		done = assert.async(),
 		node = tree.getNodeByKey("root"),
 		timer = new tools.AsyncTimer("1000 top level nodes flat with expand", 1000);
 
@@ -230,6 +238,7 @@ test("Expand 1000 top level nodes with ARIA and checkboxes (triggers expand -> r
 
 	node.setExpanded().done(function(){
 		timer.stop();
+		done();
 	});
 });
 
@@ -239,66 +248,111 @@ test("Expand 1000 top level nodes with ARIA and checkboxes (triggers expand -> r
  */
 module("Table tree");
 
-test("tabletree (6 columns): render and expand", function() {
-	expect(2);
+function _renderTable(assert, options) {
+	var node, timer, tree, $tree,
+		done = assert.async(),
+		opts = $.extend({
+			count: 1,
+			count2: 0,
+			count3: 0,
+			forceUpdate: false,
+			expandBefore: false,
+			expandAfter: false
+		}, options),
+		totalCount = opts.count * Math.max(1, opts.count2) * Math.max(1, opts.count3);
+
+	assert.expect(2);
 
 	_resetEmptyTree();
-
-//	var $ico1 = $("<img>", {src: "/src/skin-win8/icons.gif"});
-//	var $ico2 = $("<span>", {
-//		"class": "fancytree-icon"
-////		css: "/src/skin-win8/icons.gif"
-//		});
-
-	var node, timer, tree, $tree;
 
 	$tree = $("#tabletree").fancytree({
 		extensions: ["table"],
 		source: [{title: "root node", key: "root"}],
 		table: {
 			indentation: 20,  // indent 20px per node level
-			nodeColumnIdx: 0  // render the node title into the 2nd column
+			nodeColumnIdx: 1  // render the node title into the 2nd column
 		},
 		renderColumns: function(event, data) {
 			var node = data.node,
 				$tdList = $(node.tr).find(">td");
-//			$tdList.eq(0).text(node.getIndexHier()).addClass("alignRight");
-			// (index #0 is rendered by fancytree)
-//			$tdList.eq(1).append($ico2);
-//			$tdList.eq(2).append($("<img>", {src: "/src/skin-win8/icons.gif"}));
-			$tdList.eq(1).append( $("<span>", {
-				"class": "fancytree-icon"
-//					css: "/src/skin-win8/icons.gif"
-					}));
+			$tdList.eq(0).text(node.getIndexHier()).addClass("alignRight");
+//			(index #1 is rendered by fancytree)
 			$tdList.eq(2).append( $("<span>", {
-				"class": "fancytree-icon"
-//					css: "/src/skin-win8/icons.gif"
-					}));
+				text: "foo"
+				}));
 			$tdList.eq(3).append( $("<span>", {
 				"class": "fancytree-icon"
-//					css: "/src/skin-win8/icons.gif"
-					}));
-//            $tdList.eq(3).html("<input type='checkbox' name='like' value='" + node.key + "'>");
-			$tdList.eq(4).text("2013-02-17");
-			$tdList.eq(5).text("Homer Simpson");
+				}));
+			$tdList.eq(4).text("Homer Simpson");
+			$tdList.eq(5).html("<input type='text' name='cb1' value='" + node.key + "'>");
 		}
 	});
 
 	tree = $tree.fancytree("getTree");
 	node = tree.getNodeByKey("root");
-	timer = new tools.AsyncTimer("1000 nodes flat and expand", 1000);
+	timer = new tools.AsyncTimer(totalCount + " nodes", totalCount);
+	// timer = new tools.AsyncTimer(totalCount + " nodes flat and expand", totalCount);
 //    var timer = new AsyncTimer("1000 nodes (10 x 10 x 10) and force render(deep=true)");
 
-	addNodes(node, 1000, 0, 0);
-//    addNodes(node, 10, 10, 10);
-//  addNodes(node, 1, 390, 0);
+	if( opts.expandBefore ) {
+		node.addChildren({title: "dummy (to make root expandable)"});
+		node.setExpanded();
+	}
+	addNodes(node, opts.count, opts.count2, opts.count3, opts.forceUpdate);
+
 	timer.subtime("addNodes");
 
 //    tree.render(true, true);
 //    timer.subtime("render");
 
-	node.setExpanded().done(function(){
+	if( opts.expandAfter ) {
+		node.setExpanded().done(function(){
+			timer.stop();
+			done();
+		});
+	} else {
 		timer.stop();
+		done();
+	}
+}
+
+
+test("tabletree (6 columns): render collapsed", function(assert) {
+	_renderTable(assert, {
+		count: 1000,
+		forceUpdate: true,
+		expandBefore: false,
+		expandAfter: false
+	});
+});
+
+
+test("tabletree (6 columns): render, then expand", function(assert) {
+	_renderTable(assert, {
+		count: 1000,
+		forceUpdate: false,
+		expandBefore: false,
+		expandAfter: true
+	});
+});
+
+
+test("tabletree (6 columns): render while expanded", function(assert) {
+	_renderTable(assert, {
+		count: 100,
+		forceUpdate: true,
+		expandBefore: true,
+		expandAfter: false
+	});
+});
+
+
+test("tabletree (6 columns): render while expanded with enableUpdate(false)", function(assert) {
+	_renderTable(assert, {
+		count: 100,
+		forceUpdate: false,
+		expandBefore: true,
+		expandAfter: false
 	});
 });